fix all tests and warnings
